--- a/packages/api/app-prod.yaml
+++ b/packages/api/app-prod.yaml
@@ -1,5 +1,9 @@
+service: production
 runtime: custom
 env: flex
 
 manual_scaling:
   instances: 2
+
+resources:
+  memory_gb: 4