--- a/packages/api/app-programize.yaml
+++ b/packages/api/app-programize.yaml
@@ -4,3 +4,6 @@ env: flex
 
 manual_scaling:
   instances: 1
+
+resources:
+  memory_gb: 2