--- a/packages/api/src/utils/sofar.test.ts
+++ b/packages/api/src/utils/sofar.test.ts
@@ -67,7 +67,7 @@ test('it process Sofar Wave Date API for surface temperature', async () => {
   const yesterday = yesterdayDate.toISOString();
 
   const response = await sofarWaveData(
-    'SPOT-0792',
+    'SPOT-30339R',
     process.env.SOFAR_API_TOKEN,
     yesterday,
     today,