## Issue
Upload sheet data flow is slow and process time grows exponentially as the size of the file grows.

Some benchmarks using upload-sonde-script:
|# Rows | Grouping |Converting|Total|
|---|---|---|---|
|10k|0s|3s|47s|
|100k|90s|5m|7m|
|200k|5m|26m|32m|

## Changes
- Remove array spreading for adding new elements. This resulted in allocating and copying the whole array on each iteration. Use `.push()` or pre allocate the array and fill in the elements.
- Remove constant calculations out of the main loop.

|# Rows | Grouping |Converting|Total|
|---|---|---|---|
|10k|0s|0s|28s|
|100k|0s|0.3s|30s|
|200k|1s|0.5s|30s|